requirements.txt
Django~=4.0.7
djangorestframework~=3.12.4
python-dotenv~=0.19.0
psycopg2-binary~=2.9.3
faker-e164~=0.1.0
Faker~=8.12.1
factory-boy~=3.2.0
pytest~=6.2.4
pytest-django~=4.4.0
requests~=2.26.0
drf-yasg~=1.20.0
black~=21.7b0
Pillow~=8.3.2
imgpy~=1.1.0
jsonschema~=3.2.0
django-filter~=2.4.0
django-cors-headers~=3.9.0
drf-nested-routers~=0.93.3
lxml~=4.9.1
openpyxl~=3.0.9
django-storages~=1.12.3
botocore~=1.23.38
boto3~=1.20.38
django-extensions~=3.1.5
django-oauth-toolkit~=1.7.0
gunicorn==20.0.4
django-stdimage~=5.3.0
celery~=5.2.7
redis~=3.5.3
whitenoise==5.2.0
sentry-sdk==1.9.8
yfinance==0.2.13
pandas==1.5.3
numpy==1.23.5
django-celery-results==2.5.0
gevent==22.10.2
scikit-learn==1.2.2
keras==2.12.0
tensorflow==2.12.0
optuna==3.1.1
pandas-ta==0.3.14b